refactor(parser): extract small literal extractors into types/literals
Plan 2, step 4 of N. Move `literal_str` / `literal_str_list` /
`literal_u32` into `crates/toolr-core/src/parser/types/literals.rs`
(32 lines). All three were private helpers used by the not-yet-
extracted `extract_arg_metadata` (step 5) and `parse_inline_arg_section`.

Visibility promoted to `pub(super)` so callers in sibling modules
(future `arg_metadata.rs`, `resolve.rs`) can reach them.

`types/mod.rs` shrinks 944 → 920 lines.

Tests: 24 / 24 parser::types::tests still pass.
